Move files around into ViewModels and Models new directories
