+++ QuoteDroplet/Models/Quote.swift
