+++ QuoteDroplet/ViewModels/AuthorHelper.swift
