+++ QuoteDroplet/ViewModels/JSONHelper.swift
@@ -11,26 +11,6 @@ import UserNotifications
 import UIKit
 import Foundation
 
-struct QuoteJSON: Codable {
-    let id: Int
-    var text: String
-    var author: String
-    let classification: String
-}
-
-
-extension Quote {
-    func toQuoteJSON() -> QuoteJSON {
-        return QuoteJSON(id: self.id, text: self.text, author: self.author ?? "", classification: self.classification ?? "")
-    }
-}
-
-extension QuoteJSON {
-    func toQuote() -> Quote {
-        return Quote(id: self.id, text: self.text, author: self.author, classification: self.classification, likes: 0)
-    }
-}
-
 func loadQuotesFromJSON() -> [QuoteJSON] {
     // Load quotes from JSON file
     guard let path = Bundle.main.path(forResource: "QuotesBackup", ofType: "json") else {
