+++ QuoteDroplet/ViewModels/LocalQuotesHelper.swift
