+++ QuoteDroplet/ViewModels/QuoteBox.swift
