+++ QuoteDroplet/Views/Helpers/ColorHelper.swift
