+++ QuoteDroplet/Views/Helpers/FontHelper.swift
@@ -1,0 +1,15 @@
+//
+//  FontHelper.swift
+//  Quote Droplet
+//
+//  Created by Daniel Agapov on 2024-10-23.
+//
+
+import Foundation
+
+// Fonts for widget and widget preview
+let availableFonts = [
+    "Georgia", "Times New Roman",
+    "Verdana", "Palatino", "Baskerville", "Didot",
+    "Optima", "Arial"
+]
